feat: Refactor data import handling by creating DataImportController and DataImportService; remove legacy dataImportController
